src/robot_perception/requirements.txt
pydantic>=2.7
numpy>=1.26
opencv-python>=4.9
open3d>=0.19
ultralytics>=8.3
pupil-apriltags>=1.0.4
PyYAML>=6.0
transforms3d>=0.4
